+++ src/theme/delete.php
@@ -48,4 +48,4 @@ outputBanner($wiki);
     </nav>
   </div>
 </section>
-<?php outputFooter($wiki); ?>
+<?php outputFooter($wiki, $config); ?>