+++ src/theme/login.php
@@ -37,4 +37,4 @@ outputBanner($wiki);
     </div>
   </div>
 </section>
-<?php outputFooter($wiki); ?>
+<?php outputFooter($wiki, $config); ?>